+++ game/missiongenerator/flotgenerator.py
@@ -215,7 +215,7 @@ class FlotGenerator:
                         u = random.choices(
                             manpads, weights=[m.spawn_weight for m in manpads]
                         )[0]
-                        self.mission.vehicle_group(
+                        infantry_group = self.mission.vehicle_group(
                             side,
                             namegen.next_infantry_name(side, u),
                             u.dcs_unit_type,
@@ -224,6 +224,7 @@ class FlotGenerator:
                             heading=forward_heading.degrees,
                             move_formation=PointAction.OffRoad,
                         )
+                        infantry_group.hidden_on_mfd = True
             return
 
         possible_infantry_units = set(faction.infantry_with_class(UnitClass.INFANTRY))
@@ -240,7 +241,7 @@ class FlotGenerator:
             weights=[u.spawn_weight for u in infantry_choices],
             k=INFANTRY_GROUP_SIZE,
         )
-        self.mission.vehicle_group(
+        infantry_group = self.mission.vehicle_group(
             side,
             namegen.next_infantry_name(side, units[0]),
             units[0].dcs_unit_type,
@@ -249,10 +250,11 @@ class FlotGenerator:
             heading=forward_heading.degrees,
             move_formation=PointAction.OffRoad,
         )
+        infantry_group.hidden_on_mfd = True
 
         for unit in units[1:]:
             position = infantry_position.random_point_within(55, 5)
-            self.mission.vehicle_group(
+            infantry_group = self.mission.vehicle_group(
                 side,
                 namegen.next_infantry_name(side, unit),
                 unit.dcs_unit_type,
@@ -261,6 +263,7 @@ class FlotGenerator:
                 heading=forward_heading.degrees,
                 move_formation=PointAction.OffRoad,
             )
+            infantry_group.hidden_on_mfd = True
 
     def _set_reform_waypoint(
         self, dcs_group: VehicleGroup, forward_heading: Heading
@@ -771,6 +774,7 @@ class FlotGenerator:
             group_size=count,
             heading=heading.degrees,
         )
+        group.hidden_on_mfd = True
 
         self.unit_map.add_front_line_units(group, cp, unit_type)